Fixes requests from anonymous user
When the restrictor depends on db-values of the request user, then it has to pay attanchen, that the request is not send
for the anonsmous user (userID==0) since the anonyomus user is not in the database.

There were two cases, where a check was missing. This PR adds these checks.

Fixes #789